+++ CHANGELOG.md
@@ -13,6 +13,7 @@ All notable changes to this project will be documented in this file.
 - **NEW**: Can switch monitors via keyboard
 - **NEW**: Can move windows between monitors via keyboard
 - **NEW**: Can click on a window to set focus to it
+- **NEW**: Window swallowing
 - **CHANGE**: Renamed `focus_previous` to `focus_prev`
 - **CHANGE**: Invalid sample config
 - **CHANGE**: Parser `$HOME` searching order. XDG Compliance